libs/py/octo_render/pyproject.toml
[tool.poetry]
name = "octo-render"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{include = "octo_render"}]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
fastapi = "^0.104.1"
selenium = "^4.15.2"
pyvirtualdisplay = "^3.0"
pillow = "^10.1.0"

[tool.poetry.group.rpi]
optional = true

[tool.poetry.group.rpi.dependencies]
inky = {extras = ["example-depends", "rpi"], version = "^1.5.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
